# docs/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "不動産投資意思決定支援 API",
        "title": "Yield-Guard API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/area-discovery": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "area"
                ],
                "summary": "投資エリア探索",
                "parameters": [
                    {
                        "type": "string",
                        "description": "都道府県コード (例: 13)",
                        "name": "prefecture",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "予算 (円)",
                        "name": "budget",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "目標利回り (例: 0.07)",
                        "name": "yield",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.AreaDiscoveryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/geocode": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "ジオコーディング",
                "parameters": [
                    {
                        "type": "string",
                        "description": "住所文字列",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GeocodeResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/renovation/analyze": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "renovation"
                ],
                "summary": "リフォームROI分析",
                "parameters": [
                    {
                        "description": "リフォーム分析リクエスト",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RenovationInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.RenovationResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.GeocodeResult": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "locationType": {
                    "type": "string"
                }
            }
        },
...
        "domain.AreaDiscoveryItem": {
            "type": "object",
            "properties": {
                "dataSufficient": {
                    "description": "取引件数3件以上",
                    "type": "boolean"
                },
                "landPriceTrend": {
                    "description": "\"上昇\" | \"安定\" | \"下落\" | \"不明\"",
                    "type": "string"
                },
                "medianTsubo": {
                    "description": "坪単価中央値（円）",
                    "type": "number"
                },
                "municipalityCode": {
                    "type": "string"
                },
                "municipalityName": {
                    "type": "string"
                },
                "transactionCount": {
                    "description": "取引件数",
                    "type": "integer"
                },
                "yieldDifficulty": {
                    "description": "\"achievable\" | \"slightly-difficult\" | \"difficult\"",
                    "type": "string"
                },
                "yieldDifficultyLabel": {
                    "description": "日本語ラベル",
                    "type": "string"
                }
            }
        },
        "domain.AreaDiscoveryResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.AreaDiscoveryItem"
                    }
                },
                "prefecture": {
                    "type": "string"
                }
            }
        },
...
        "domain.ClassifiedRenovationItem": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "expectedMonthlyRentIncrease": {
                    "type": "number"
                },
                "isCapitalExpenditure": {
                    "type": "boolean"
                },
                "isSelfWork": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "selfLaborHours": {
                    "type": "number"
                },
                "virtualLaborCost": {
                    "type": "number"
                }
            }
        },
...
        "domain.RenovationInput": {
            "type": "object",
            "properties": {
                "annualBaseRent": {
                    "type": "number"
                },
                "annualExpenses": {
                    "type": "number"
                },
                "effectiveTaxRate": {
                    "type": "number"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.RenovationItem"
                    }
                },
                "propertyPrice": {
                    "type": "number"
                },
                "selfLaborRatePerHour": {
                    "type": "number"
                }
            }
        },
        "domain.RenovationItem": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "expectedMonthlyRentIncrease": {
                    "type": "number"
                },
                "isSelfWork": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "selfLaborHours": {
                    "type": "number"
                }
            }
        },
        "domain.RenovationResult": {
            "type": "object",
            "properties": {
                "actualYield": {
                    "type": "number"
                },
                "annualRentIncrease": {
                    "type": "number"
                },
                "capitalExpenditures": {
                    "type": "number"
                },
                "classifiedItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ClassifiedRenovationItem"
                    }
                },
                "isRecoverable": {
                    "description": "家賃アップがある場合のみ true",
                    "type": "boolean"
                },
                "recoveryYears": {
                    "type": "number"
                },
                "repairExpenses": {
                    "type": "number"
                },
                "taxSavings": {
                    "type": "number"
                },
                "totalRenovationCost": {
                    "type": "number"
                },
                "virtualLaborCost": {
                    "type": "number"
                }
            }
        },
...
    }
}